move `controllers` to `internal/controllers`
