+++ .gitignore
@@ -26,3 +26,6 @@ bin
 
 build/
 data
+
+# CRDs for fuzzing tests.
+internal/controllers/testdata/crd
