+++ internal/controllers/scan_test.go
