+++ main.go
@@ -45,7 +45,7 @@ import (
 	// +kubebuilder:scaffold:imports
 
 	imagev1 "github.com/fluxcd/image-reflector-controller/api/v1beta2"
-	"github.com/fluxcd/image-reflector-controller/controllers"
+	"github.com/fluxcd/image-reflector-controller/internal/controllers"
 	"github.com/fluxcd/image-reflector-controller/internal/database"
 	"github.com/fluxcd/image-reflector-controller/internal/features"
 )
